Allows to show the load order and list plugins.
i don't think allowing to edit the load order un the CLI will be easy.
maybe it's doable with a TUI ? but if i continue to work with command line arguments it doesn't seem possible to do it cleanly.
maybe i can do the same thing i did with the mod order. numbering mods and assking two numbers to swap. but this is shit.

i only see 2 way this might go forward.

1. a gtk based ui like the one being developped in the interface branch but i doubt it will ever be finished if i don't intervene at some point.

2. a tui im really curious about tuis like nmtui or "n"(a node version manager).